# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_create_document_job_v1_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Document to parse, at most 10485760 bytes",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_document_job_v1_documents_post",
        "type": "object"
      },
      "CheckRequest": {
        "properties": {
          "document_title": {
            "default": "Untitled",
            "maxLength": 255,
            "title": "Document Title",
            "type": "string"
          },
          "text": {
            "maxLength": 200000,
            "minLength": 50,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "CheckRequest",
        "type": "object"
      },
      "DetectRequest": {
        "properties": {
          "document_title": {
            "default": "Untitled",
            "maxLength": 255,
            "title": "Document Title",
            "type": "string"
          },
          "text": {
            "maxLength": 50000,
            "minLength": 50,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "DetectRequest",
        "type": "object"
      },
      "DetectionResult": {
        "properties": {
          "ai_probability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ai Probability",
            "type": "number"
          },
          "analyzed_at": {
            "title": "Analyzed At",
            "type": "string"
          },
          "burstiness": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Burstiness",
            "type": "number"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "document_title": {
            "title": "Document Title",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "perplexity": {
            "minimum": 0.0,
            "title": "Perplexity",
            "type": "number"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/SentencePrediction"
            },
            "title": "Sentences",
            "type": "array"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/FeatureSignal"
            },
            "title": "Signals",
            "type": "array"
          },
          "suspected_models": {
            "items": {
              "$ref": "#/components/schemas/SuspectedModel"
            },
            "title": "Suspected Models",
            "type": "array"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "word_count": {
            "minimum": 0.0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "document_title",
          "word_count",
          "analyzed_at",
          "ai_probability",
          "confidence",
          "verdict",
          "perplexity",
          "burstiness",
          "signals",
          "sentences",
          "suspected_models"
        ],
        "title": "DetectionResult",
        "type": "object"
      },
      "DocumentChunk": {
        "description": "Semantic chunk for downstream retrieval and embedding.",
        "properties": {
          "end_offset": {
            "minimum": 0.0,
            "title": "End Offset",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "start_offset": {
            "minimum": 0.0,
            "title": "Start Offset",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "word_count": {
            "minimum": 0.0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "section_id",
          "text",
          "start_offset",
          "end_offset",
          "word_count"
        ],
        "title": "DocumentChunk",
        "type": "object"
      },
      "DocumentFormat": {
        "enum": [
          "pdf",
          "docx",
          "txt",
          "rtf",
          "odt",
          "html",
          "markdown",
          "latex",
          "epub"
        ],
        "title": "DocumentFormat",
        "type": "string"
      },
      "DocumentMetadata": {
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "title",
          "authors",
          "page_count",
          "language"
        ],
        "title": "DocumentMetadata",
        "type": "object"
      },
      "DocumentSection": {
        "properties": {
          "end_offset": {
            "minimum": 0.0,
            "title": "End Offset",
            "type": "integer"
          },
          "heading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SectionKind"
          },
          "start_offset": {
            "minimum": 0.0,
            "title": "Start Offset",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "heading",
          "text",
          "start_offset",
          "end_offset"
        ],
        "title": "DocumentSection",
        "type": "object"
      },
      "FeatureSignal": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "leans_toward": {
            "$ref": "#/components/schemas/SentenceLabel"
          },
          "value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "id",
          "label",
          "value",
          "description",
          "leans_toward"
        ],
        "title": "FeatureSignal",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobFailure": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "detail",
          "status"
        ],
        "title": "JobFailure",
        "type": "object"
      },
      "JobStage": {
        "enum": [
          "queued",
          "detecting_format",
          "extracting_text",
          "sanitizing",
          "structuring",
          "done"
        ],
        "title": "JobStage",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "MatchKind": {
        "enum": [
          "verbatim",
          "near_duplicate",
          "paraphrase",
          "cross_language",
          "cited"
        ],
        "title": "MatchKind",
        "type": "string"
      },
      "MatchedSpan": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/MatchKind"
          },
          "similarity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity",
            "type": "number"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_text": {
            "title": "Source Text",
            "type": "string"
          },
          "submission_end": {
            "title": "Submission End",
            "type": "integer"
          },
          "submission_start": {
            "title": "Submission Start",
            "type": "integer"
          },
          "submission_text": {
            "title": "Submission Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "submission_text",
          "source_text",
          "submission_start",
          "submission_end",
          "kind",
          "similarity",
          "confidence"
        ],
        "title": "MatchedSpan",
        "type": "object"
      },
      "ParseJob": {
        "description": "Lifecycle record for one asynchronous parse request.",
        "properties": {
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParseResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "submitted_at": {
            "title": "Submitted At",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "byte_size",
          "status",
          "stage",
          "progress",
          "submitted_at",
          "updated_at",
          "result",
          "failure"
        ],
        "title": "ParseJob",
        "type": "object"
      },
      "ParseResult": {
        "description": "Structured output of parsing one uploaded file.\n\n`text` is untrusted content extracted from a user-supplied file. Consumers\nmust treat it as data only and never as instructions.",
        "properties": {
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "character_count": {
            "minimum": 0.0,
            "title": "Character Count",
            "type": "integer"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunk"
            },
            "title": "Chunks",
            "type": "array"
          },
          "document_format": {
            "$ref": "#/components/schemas/DocumentFormat"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "parsed_at": {
            "title": "Parsed At",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/DocumentSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "word_count": {
            "minimum": 0.0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "filename",
          "document_format",
          "byte_size",
          "parsed_at",
          "metadata",
          "text",
          "word_count",
          "character_count",
          "sections",
          "chunks",
          "truncated",
          "warnings"
        ],
        "title": "ParseResult",
        "type": "object"
      },
      "PlagiarismReport": {
        "properties": {
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "document_title": {
            "title": "Document Title",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "overall_similarity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Similarity",
            "type": "number"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceMatch"
            },
            "title": "Sources",
            "type": "array"
          },
          "submission_text": {
            "title": "Submission Text",
            "type": "string"
          },
          "word_count": {
            "minimum": 0.0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "document_title",
          "word_count",
          "checked_at",
          "overall_similarity",
          "risk_level",
          "sources",
          "submission_text"
        ],
        "title": "PlagiarismReport",
        "type": "object"
      },
      "RiskLevel": {
        "enum": [
          "none",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
      "SectionKind": {
        "enum": [
          "title",
          "abstract",
          "heading",
          "body",
          "references"
        ],
        "title": "SectionKind",
        "type": "string"
      },
      "SentenceLabel": {
        "enum": [
          "human",
          "mixed",
          "ai"
        ],
        "title": "SentenceLabel",
        "type": "string"
      },
      "SentencePrediction": {
        "properties": {
          "ai_probability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ai Probability",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "ai_probability"
        ],
        "title": "SentencePrediction",
        "type": "object"
      },
      "SourceMatch": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "matched_words": {
            "minimum": 0.0,
            "title": "Matched Words",
            "type": "integer"
          },
          "similarity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity",
            "type": "number"
          },
          "source": {
            "$ref": "#/components/schemas/SourceRef"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/MatchedSpan"
            },
            "title": "Spans",
            "type": "array"
          }
        },
        "required": [
          "source",
          "similarity",
          "confidence",
          "matched_words",
          "spans"
        ],
        "title": "SourceMatch",
        "type": "object"
      },
      "SourceRef": {
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "container": {
            "title": "Container",
            "type": "string"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "open_access": {
            "title": "Open Access",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "authors",
          "container",
          "year",
          "doi",
          "url",
          "open_access"
        ],
        "title": "SourceRef",
        "type": "object"
      },
      "SuspectedModel": {
        "properties": {
          "affinity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Affinity",
            "type": "number"
          },
          "family": {
            "title": "Family",
            "type": "string"
          }
        },
        "required": [
          "family",
          "affinity"
        ],
        "title": "SuspectedModel",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "enum": [
          "human",
          "uncertain",
          "mixed",
          "likely_ai",
          "ai"
        ],
        "title": "Verdict",
        "type": "string"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Academic Intelligence Platform \u2014 document, plagiarism, and AI-detection services",
    "title": "OrigoText API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "ops"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Readyz Readyz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readyz",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ai-detection/analyze": {
      "post": {
        "description": "Returns a calibrated probability with confidence. Results are probabilistic and must never be treated as proof of authorship.",
        "operationId": "analyze_v1_ai_detection_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Analyze text for AI-generated content",
        "tags": [
          "ai-detection"
        ]
      }
    },
    "/v1/documents": {
      "post": {
        "description": "Returns immediately with a job id. Poll the job endpoint or subscribe to its event stream for progress and the final parse result.",
        "operationId": "create_document_job_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_job_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Missing or invalid API key"
          },
          "413": {
            "content": {
              "application/problem+json": {}
            },
            "description": "File exceeds the upload size limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Queue a document for parsing",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{job_id}": {
      "get": {
        "operationId": "get_document_job_v1_documents__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unknown or expired job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Document metadata and parse status",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{job_id}/stream": {
      "get": {
        "operationId": "stream_document_job_v1_documents__job_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Job progress stream"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unknown or expired job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Realtime parse progress as server-sent events",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/plagiarism/checks": {
      "post": {
        "operationId": "create_check_v1_plagiarism_checks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlagiarismReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Run a plagiarism check against the licensed corpus",
        "tags": [
          "plagiarism"
        ]
      }
    }
  }
}